+++ src/main/java/me/ajlane/geo/repast/succession/pathway/LcsTransitionMapReader.java
@@ -3,7 +3,7 @@ package me.ajlane.geo.repast.succession.pathway;
 import me.ajlane.geo.repast.succession.pathway.aliased.AliasedLcsTransitionMap;
 import me.ajlane.geo.repast.succession.pathway.coded.CodedLcsTransitionMap;
 
-public interface LcsTransitionMapFactory {
+public interface LcsTransitionMapReader {
   CodedLcsTransitionMap getCodedLcsTransitionMap();
   AliasedLcsTransitionMap getAliasedLcsTransitionMap();
 }